# onboarding-app/openapi.yaml
openapi: 3.0.3
info:
  title: Employee Onboarding App
  description: >
    REST API for managing employee onboarding processes and company car fleet operations.
    Supports creating employee records, tracking CV creation and car provisioning status,
    assigning cars from the fleet, and managing car availability.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/car/{id}:
    get:
      tags:
        - Car Fleet Management
      summary: Get car details
      description: >
        Retrieves detailed information about a specific car including its name (make/model),
        license plate number, availability status, assigned employee information (if any),
        price range class, and current operational status. Price range classes:
        1=Economy, 2=Mid-range, 3=Premium, 4=Luxury.
      operationId: getCarById
      parameters:
        - $ref: '#/components/parameters/CarId'
      responses:
        '200':
          description: Car details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '404':
          description: Car not found with the specified ID

    post:
      tags:
        - Car Fleet Management
      summary: Update car status
      description: >
        Updates the operational status of a specific car. Allowed status values are:
        'Available' (ready for assignment), 'In Maintenance' (undergoing repairs or service),
        and 'Reserved' (held for future assignment). Setting status to 'Available' is only
        permitted if the car is not currently assigned to an employee.
      operationId: updateCarStatus
      parameters:
        - $ref: '#/components/parameters/CarId'
        - name: status
          in: query
          description: New status for the car
          required: true
          schema:
            type: string
            enum:
              - Available
              - In Maintenance
              - Reserved
            example: In Maintenance
      responses:
        '200':
          description: Car status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '400':
          description: >
            Invalid status value provided – must be one of: Available, In Maintenance, Reserved.
            Setting Available is not allowed when the car is assigned to an employee.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Car not found with the specified ID
...
components:
  parameters:
...
    CarId:
      name: id
      in: path
      description: Unique identifier of the car
      required: true
      schema:
        type: string
        example: car-123e4567
  schemas:
...
    Car:
      type: object
      description: Car entity with availability and assignment information
      properties:
        id:
          type: string
          description: Unique identifier for the car
          example: car-123e4567
        name:
          type: string
          description: Car make and model
          example: BMW 3 Series
        licensePlate:
          type: string
          description: Car license plate number
          example: 2-JKL-012
        available:
          type: boolean
          description: Whether the car is currently available for assignment
          default: true
          example: true
        assignedEmployeeId:
          type: string
          description: ID of the employee assigned to this car
          example: 123e4567-e89b-12d3-a456-426614174000
        assignedEmployeeName:
          type: string
          description: Name of the employee assigned to this car
          example: John Doe
        priceRange:
          type: integer
          description: Car price range class (1=Economy, 2=Mid-range, 3=Premium, 4=Luxury)
          minimum: 1
          maximum: 4
          example: 2
        status:
          type: string
          description: Current operational status of the car
          enum:
            - Available
            - Assigned
            - In Maintenance
            - Reserved
          example: Available
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid status 'Assigned'. Allowed values: Available, In Maintenance, Reserved"